fix: start and managers handling are now in arke and do not depends on macro defined arke
