+++ lib/arke/boundary/group_manager.ex
@@ -77,8 +77,11 @@ defmodule Arke.Boundary.GroupManager do
 
     Enum.reduce(group_keys, [], fn {g, _}, groups ->
       group = get(g, project)
-      arke_id in Enum.map(group.data.arke_list, fn a -> a.id end)
+      if arke_id in Enum.map(group.data.arke_list, fn a -> a.id end) do
       [group | groups]
+      else
+      groups
+      end
     end)
   end
