+++ lib/registry/arke.json
@@ -514,7 +514,6 @@
   },
   {
     "id": "binary",
-    
       "label": "Binary",
 "parameters": [
       {
